cambios tarde 06/10/2024

+++ Moviles/My Application/app/src/main/java/com/example/myapplication/domain/model/Book.kt
@@ -4,7 +4,7 @@ package com.example.myapplication.domain.model
 data class Book(
     val name : String = "",
     val author : String = "Anonymus",
-    val score : Float = 0f,
+    val score : Int = 0,
     val releaseDate : Int = 0,
 ) {
 }
